fix redeployment routine

+++ cluster_orchestrator/cluster-manager/clients/job_management.py
@@ -111,16 +111,19 @@ def update_status(job_id, instance_number, status, status_detail=None):
     if job is None:
         return
 
-    # don't update job to running
-    if status != DeploymentStatus.RUNNING.value:
-        job["status"] = status
-
     if job.get("instance_list") is not None:
         for instance in job.get("instance_list"):
             if instance["instance_number"] == instance_number:
                 instance["status"] = status
                 if status_detail is not None:
                     instance["status_detail"] = status_detail
+
+    # Update job-level status, but only set RUNNING once all instances are running
+    if status != DeploymentStatus.RUNNING.value:
+        job["status"] = status
+    elif all(i["status"] == DeploymentStatus.RUNNING.value
+             for i in job.get("instance_list", [])):
+        job["status"] = status
 
     job_operations.update_job(job_id, job)
 
@@ -234,7 +237,7 @@ def delete_job_instance(job_id: int, instance_number: int, erase: bool = True):
                     f"Deleted instance {instance['instance_number']} of job {job_id} from DB"
                 )
 
-    if len(instance_list) <= deleted_job:
+    if erase and len(instance_list) <= deleted_job:
         job_operations.delete_job(job_id)
         logger.info(f"Deleted job {job_id} from DB as all instances were removed")
         return {}
